refactor(history-service): move definition factory to index.ts
